backend/requirements.txt
fastapi
uvicorn

pydantic
python-dotenv

bm25
requests
sarvam

python-multipart